+++ packages/ui-core/gulpfile.js
@@ -29,6 +29,7 @@ const indexTs = path.join(srcPath, 'index.ts');
 const baseLessSrc = path.join(srcPath, 'styles', 'base.less');
 const baseLessPackagePath = path.join(__dirname, 'styles');
 
+const doczReg = 'src/**/*.docz.{tsx,ts}';
 const testsReg = 'src/**/*.test.{tsx,ts}';
 const iconsReg = 'src/**/Icons.{tsx,ts}';
 
@@ -51,23 +52,22 @@ const tsConfig = {
 const babelPlugins = [
     require.resolve('@babel/plugin-transform-object-assign'),
     require.resolve('@babel/plugin-proposal-class-properties'),
-    require.resolve('@babel/plugin-proposal-object-rest-spread'),
-    require.resolve('@babel/plugin-transform-runtime'),
-    require.resolve('@babel/plugin-transform-classes'),
-    require.resolve('@babel/plugin-transform-block-scoping')
+    require.resolve('@babel/plugin-transform-runtime')
 ];
 const babelPresets = [
     '@babel/react',
-    ['@babel/env', {
-        useBuiltIns: 'entry'
+    ['@babel/preset-env', {
+        useBuiltIns: 'entry',
+        corejs: '3.6'
     }]
 ];
 const babelEsConfig = {
     presets: [
         '@babel/react',
-        ['@babel/env', {
+        ['@babel/preset-env', {
             modules: false,
-            'useBuiltIns': 'entry'
+            'useBuiltIns': 'entry',
+            corejs: '3.6'
         }]
     ],
     plugins: [
@@ -83,8 +83,9 @@ const babelEsConfig = {
 };
 const babelLibConfig = {
     presets: [
-        ['@babel/env', {
-            useBuiltIns: 'entry'
+        ['@babel/preset-env', {
+            useBuiltIns: 'entry',
+            corejs: '3.6'
         }]
     ]
 };
@@ -111,7 +112,7 @@ gulp.task('less:compile', () => {
                 .replace(/icons\//g, '')
                 .replace(/\//g, '-');
 
-            return `../icons/${newPath}`;
+            return `dist/icons/${newPath}`;
         }))
         .pipe(gulpLess(lessConfig))
         .pipe(dest(esPath))
@@ -126,8 +127,13 @@ gulp.task('less:copy-base', function() {
 gulp.task('less', gulp.series('less:compile', 'less:copy-base'));
 
 gulp.task('ts', () => {
-    const result = gulp.src(['src/**/*.{tsx,ts}', `!${testsReg}`, `!${iconsReg}`, './typings/*.d.ts'])
-        .pipe(ts(tsConfig));
+    const result = gulp.src([
+        'src/**/*.{tsx,ts}',
+        `!${doczReg}`,
+        `!${testsReg}`,
+        `!${iconsReg}`,
+        './typings/*.d.ts'
+    ]).pipe(ts(tsConfig));
 
     return merge(
         result.dts
@@ -144,7 +150,7 @@ gulp.task('ts', () => {
 });
 
 gulp.task('main', done => {
-    const components = glob.sync('src/components/**/*.{tsx,ts}', { ignore: testsReg });
+    const components = glob.sync('src/components/**/*.{tsx,ts}', { ignore: [testsReg, doczReg] });
     const utils = glob.sync('src/utils/*.ts', { ignore: testsReg });
     fs.writeFile(indexTs, generateIndex([...components, ...utils]), done);
 });