+++ packages/ui-core/src/components/Grid/GridColumn.test.tsx
@@ -8,7 +8,7 @@ describe('<GridColumn />', () => {
         expect(wrapper).toMatchSnapshot();
     });
 
-    it('it renders GridColumn with props', () => {
+    it('it renders GridColumn with props on all screens', () => {
         const wrapper = shallow(
             <GridColumn
                 wide="8"
@@ -17,6 +17,35 @@ describe('<GridColumn />', () => {
                 mobile="1"
                 all="10"
                 align="right"
+                orderAll="2"
+                offsetAll="4"
+                grow
+                flex
+            >
+                item
+            </GridColumn>
+        );
+        expect(wrapper).toMatchSnapshot();
+    });
+
+    it('it renders GridColumn with props on different screens', () => {
+        const wrapper = shallow(
+            <GridColumn
+                className="test-class"
+                wide="8"
+                desktop="4"
+                tablet="3"
+                mobile="1"
+                all="10"
+                align="right"
+                orderWide="2"
+                orderDesktop="1"
+                orderTablet="3"
+                orderMobile="4"
+                offsetWide="8"
+                offsetDesktop="7"
+                offsetTablet="6"
+                offsetMobile="5"
                 grow
                 flex
             >