+++ packages/ui-core/src/components/Grid/GridColumn.tsx
@@ -3,17 +3,43 @@ import * as PropTypes from 'prop-types';
 import './GridColumn.less';
 import cnCreate from 'utils/cn';
 
+const GridSizeValues = PropTypes.oneOf(['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12']);
+export type TGridSizeValues = '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '10' | '11' | '12';
+
 export interface IGridColumn {
     /** Size of columns on wide screens */
-    wide?: '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '10' | '11' | '12';
+    wide?: TGridSizeValues;
     /** Size of columns on desktop screens */
-    desktop?: '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '10' | '11' | '12';
+    desktop?: TGridSizeValues;
     /** Size of columns on tablet screens */
-    tablet?: '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '10' | '11' | '12';
+    tablet?: TGridSizeValues;
     /** Size of columns on mobile screens */
-    mobile?: '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '10' | '11' | '12';
+    mobile?: TGridSizeValues;
     /** Size of columns on all screens */
-    all?: '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '10' | '11' | '12';
+    all?: TGridSizeValues;
+
+    /** Order on all screens */
+    orderAll?: string;
+    /** Order on wide screens */
+    orderWide?: string;
+    /** Order on desktop screens */
+    orderDesktop?: string;
+    /** Order on tablet screens */
+    orderTablet?: string;
+    /** Order on mobile screens */
+    orderMobile?: string;
+
+    /** Size of offset on wide screens */
+    offsetWide?: TGridSizeValues;
+    /** Size of offset on desktop screens */
+    offsetDesktop?: TGridSizeValues;
+    /** Size of offset on tablet screens */
+    offsetTablet?: TGridSizeValues;
+    /** Size of offset on mobile screens */
+    offsetMobile?: TGridSizeValues;
+    /** Size of offset on all screens */
+    offsetAll?: TGridSizeValues;
+
     /** Custom alignment of column */
     align?: 'right' | 'left' | 'center';
     /** Column flex grow */
@@ -28,11 +54,24 @@ export interface IGridColumn {
 const cn = cnCreate('mfui-grid-column');
 class GridColumn extends React.Component<IGridColumn, {}> {
     static propTypes = {
-        wide: PropTypes.oneOf(['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12']),
-        desktop: PropTypes.oneOf(['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12']),
-        tablet: PropTypes.oneOf(['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12']),
-        mobile: PropTypes.oneOf(['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12']),
-        all: PropTypes.oneOf(['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12']),
+        wide: GridSizeValues,
+        desktop: GridSizeValues,
+        tablet: GridSizeValues,
+        mobile: GridSizeValues,
+        all: GridSizeValues,
+
+        orderAll: PropTypes.string,
+        orderWide: PropTypes.string,
+        orderDesktop: PropTypes.string,
+        orderTablet: PropTypes.string,
+        orderMobile: PropTypes.string,
+
+        offsetWide: GridSizeValues,
+        offsetDesktop: GridSizeValues,
+        offsetTablet: GridSizeValues,
+        offsetMobile: GridSizeValues,
+        offsetAll: GridSizeValues,
+
         align: PropTypes.oneOf(['right', 'left', 'center']),
         grow: PropTypes.bool,
         flex: PropTypes.bool,
@@ -48,10 +87,36 @@ class GridColumn extends React.Component<IGridColumn, {}> {
         all: '12',
         flex: false,
         grow: false,
+        orderWide: '0',
+        orderDesktop: '0',
+        orderTablet: '0',
+        orderMobile: '0',
+        orderAll: '0',
     };
 
     render() {
-        const { all, wide, desktop, tablet, className, flex, grow, align, children, mobile } = this.props;
+        const {
+            all,
+            wide,
+            desktop,
+            tablet,
+            mobile,
+            orderAll,
+            orderWide,
+            orderDesktop,
+            orderTablet,
+            orderMobile,
+            offsetAll,
+            offsetWide,
+            offsetDesktop,
+            offsetTablet,
+            offsetMobile,
+            className,
+            flex,
+            grow,
+            align,
+            children,
+        } = this.props;
 
         return (
             <div
@@ -66,6 +131,16 @@ class GridColumn extends React.Component<IGridColumn, {}> {
                         desktop,
                         tablet,
                         mobile,
+                        'all-order': orderAll,
+                        'wide-order': orderWide,
+                        'desktop-order': orderDesktop,
+                        'tablet-order': orderTablet,
+                        'mobile-order': orderMobile,
+                        'offset-all': offsetAll,
+                        'offset-wide': offsetWide,
+                        'offset-desktop': offsetDesktop,
+                        'offset-tablet': offsetTablet,
+                        'offset-mobile': offsetMobile,
                     },
                     className
                 )}