+++ packages/ui-core/src/components/Header/Header.tsx
@@ -16,7 +16,7 @@ interface IHeaderProps {
     className?: string;
     children?: JSX.Element[] | Element[] | JSX.Element | string | Element;
     /** Click handler */
-    onClick?(e: React.SyntheticEvent<EventTarget>): void;
+    onClick?: (e: React.SyntheticEvent<EventTarget>) => void;
 }
 
 const cn = cnCreate('mfui-header');
@@ -44,7 +44,7 @@ class Header extends React.Component<IHeaderProps, {}> {
     }
 
     render() {
-        const ElementType = this.props.as as string;
+        const ElementType = this.props.as as React.ElementType;
         const { color, margin, as: level, onClick } = this.props;
 
         return (