+++ packages/ui-core/src/components/List/List.tsx
@@ -38,7 +38,7 @@ class List extends React.Component<IListProps, {}> {
     };
 
     render() {
-        const ElementType = this.props.as as string;
+        const ElementType = this.props.as as React.ElementType;
         const { hAlign, color, weight, className, children } = this.props;
 
         return (