+++ packages/ui-shared/typings/custom-typings.d.ts
@@ -1,5 +1,3 @@
-declare module 'prop-types';
-
 declare module '*.svg' {
     const content: any;
     export default content;