+++ packages/ui-core/CHANGELOG.md
@@ -2,7 +2,44 @@
 
 ## Unreleased
 - **Header**
-    - correct header css for style guide
+    - correct header css for style guide, delete h6 props
+- **Carousel**
+    - update classes prop
+
+## Release 1.4.0
+- **Button**
+    - add font-color for hover button type
+- **GridColumn**
+    - add offset props
+- **ContentArea**
+    - `mobileInnerPadding` and `className` properties added
+- **Paragraph**
+    - line-height in mixin for small font fixed according to the guide
+
+## Release 1.3.4
+- **Carousel**
+    - add classes props for flexible styling of the component
+
+## Release 1.3.3
+- **Carousel**
+    - remove margin beetween slides in landing theme
+
+## Release 1.3.2
+- **Carousel**
+    - fix slick slide padding for lk themed carousel
+
+## Release 1.3.1
+- **Carousel**
+    - added top and bottom padding for overflow slide content
+
+## Release 1.3.0
+- **GridColumn**
+    - add `order` property for all screens
+
+## Release 1.2.0
+- **Carousel**
+    - `hasPaddingBetweenSlides` property added (zero padding between slides)
+    - fix showcase slider padding
 
 ## Release 1.1.0
 - **Carousel**